Update GepardEngine constructor
